Updated Catergories and NavBar
