+++ Nourish/src/components/HomePage.jsx
@@ -3,7 +3,7 @@ import React from 'react'
 const HomePage = () => {
   return (
     <div>
-      <h1>Noursih</h1>
+      <h1>Nourish</h1>
       <h2>Wellness starts here..</h2>
     </div>
   )
