+++ Nourish/src/context/NourishProvider.jsx
@@ -5,18 +5,7 @@ import { useState, useEffect } from "react";
 
 const NourishProvider = ({children}) => {
 
-    const [data, setData] = useState([]);
-
-useEffect(() => {
-    const fetchData = async () => {
-      const [data, error] = await handleFetch(`https://www.themealdb.com/api/json/v1/1/categories.php`)
-      if (data) setData(data)
-      if (error) setError(error.message)
-    }
-  fetchData();
-  }, []);
-
-
+  const [data, setData] = useState([]);
 
   return (
     <div>
